Bugfix if pongo2 template file didn't exist
If the pongo2 template didn't exist, it would call http.Error()
as expected, but then it carried on executing code past that point.

Refactored this function so that this does not happen, now if the
template does not exist, it simply prints that on the HTTP response
and then there is no panic following that as before.