Start framework from scratch by forking gocraft/web.
I liked the idea of custom a Context struct in gocraft/web,
but I also needed a more cohesive framework where the middlewares
provided by the framework work together nicely with the Context
and the Context remains strictly typed.

This meant that apps using the framework must now embed
the Context struct provided by the framework, that changes
things a bit.  A minor change was required in ServeHTTP in
order to create the embedded Context struct to make this work.

I also wanted to provide an application Config struct when
creating the root router, while the original gocraft/web
just stores things like the logger instance in a global variable,
meaning you couldn't create multiple "apps", I didn't want
to do this.

I started off wanting to write a framework "on top of" gocraft/web,
however in the end I was forced to fork it instead.